pyproject.toml
[project]
name = "llm-lab"
version = "0.1.0"
description = "A playground for experimenting with LLMs and multi-language agents."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.86.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]

[tool.setuptools]
package-dir = {"" = "python/src"}

[tool.setuptools.packages.find]
where = ["python/src"]
